added DaoFactory refactored

+++ src/main/java/com/kpi/controllers/RemoveSalad.java
@@ -1,5 +1,7 @@
 package com.kpi.controllers;
 
+import com.kpi.dao.DaoFactory;
+import com.kpi.dao.SaladDAO;
 import com.kpi.models.Salad;
 import com.kpi.service.SaladService;
 
@@ -14,16 +16,15 @@ import java.sql.SQLException;
 @WebServlet(name = "RemoveSalad", urlPatterns = "/removeSalad")
 public class RemoveSalad extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        DaoFactory daoFactory = new DaoFactory();
+        SaladDAO saladService = daoFactory.getSaladDAO();
 
-        SaladService saladService = new SaladService();
         Long saladId = Long.parseLong(request.getParameter("id"));
         try {
             Salad s = saladService.getById(saladId);
-
-            saladService.remove(s);
-        }
-        catch (SQLException e)
-        {
+            if (s != null)
+                saladService.remove(s);
+        } catch (SQLException e) {
             e.printStackTrace();
         }
